gpt_engineer/applications/web/requirements.txt
# Requirements for GPT Engineer Web IDE
gradio>=4.0.0
python-dotenv>=0.21.0
openai>=1.0.0

# Optional - Gemini support
google-generativeai>=0.3.0
